+++ packages/atti-compo/src/index.native.js
@@ -1,5 +1,5 @@
 import Button from './components/Button'
-import { Text } from './components/Typography'
+import { Text } from './components/Typographies'
 import ThemeProvider from './components/ThemeProvider'
 
 import { defaultTheme } from './themes'